requirements.txt
requests
feedparser
python-dotenv
schedule
markdown